requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
openai==1.51.0
python-jose[cryptography]==3.3.0
PyJWT==2.9.0
Pillow==10.4.0
PyMuPDF==1.24.10
jsonschema==4.23.0
google-auth==2.35.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
redis==5.1.1
python-multipart==0.0.12
